--- a/tensorflow_serving/model_servers/server.h
+++ b/tensorflow_serving/model_servers/server.h
@@ -39,6 +39,7 @@ class Server {
     //
     tensorflow::int32 grpc_port = 8500;
     tensorflow::string grpc_channel_arguments;
+    tensorflow::string grpc_socket_path;
 
     //
     // HTTP Server options.