@netfs 2018-10-29
thanks for the change! i will review this sometime this week.

do you have any benchmarks to show case improvements in latency (or throughput)
when going over unix domain sockets vs tcp/ip for ML inference? i'd be curious to see
the results.