@vtorhonen 2019-01-03
Sorry, I had too much fun with Tensorflow and it took a while to get back here. :)

Here are some benchmark results from a project that runs Tensorflow Serving on GPUs on Google Kubernetes Engine. Our client app uses [Applifier/go-tensorflow](https://github.com/applifier/go-tensorflow) to interface with Tensorflow Serving over gRPC. We also built a benchmark tool with the same library. We did two separate runs with the benchmark tool, first by calling Tensorflow Serving over a UNIX domain socket and then by calling it over the default TCP socket. There was a small pause between the runs.

This first graph shows the average rate of successful predictions per second.

![pr1159-predictions-per-second](https://user-images.githubusercontent.com/3088277/50639511-665f9180-0f6a-11e9-877c-962630535bdc.png)

The second graph shows the average endpoint latency seen by the gRPC client.

![pr1159-grpc-endpoint-latency](https://user-images.githubusercontent.com/3088277/50639615-c3f3de00-0f6a-11e9-8301-6d9f9183a9a7.png)

There's a huge difference in latency with the default TCP socket, which is actually why our test run eventually failed to finish. Peak median latency increased from 30 ms to 135 ms, where p99 latency increased from 260 ms to 480 ms.

Unfortunately I don't have a pathological and reproducible example about this.