@netfs 2019-01-22
> @netfs Do you still need more info?

nope. this looks good. thanks for the change, started to take a look/review!