@ndeepesh 2022-04-28
Hi @vtorhonen 
Do you have the code for the benchmark that you performed? I am trying a similar thing and latency (p99) is worse for unix domain socket. Just wanted to understand if you ran your configuration with some specific configurations